+++ packages/admin/src/components/NewArticleModal/index.jsx
@@ -9,6 +9,7 @@ import { Button, Modal } from 'antd';
 import moment from 'moment';
 import { stopMenuKeydown } from '@/services/van-blog/editableKeyboard';
 import AuthorField from '../AuthorField';
+import CoverImageField from '../CoverImageField';
 import PathnameField from '../PathnameField';
 import TagSelectField from '../TagSelectField';
 
@@ -157,6 +158,7 @@ export default function (props) {
         tooltip="设置后会替换掉文章页底部默认的版权声明文字，留空则根据系统设置中的相关选项进行展示"
         placeholder="设置后会替换掉文章底部默认的版权"
       />
+      <CoverImageField name="coverC" id="coverC" />
       </div>
     </ModalForm>
   );