+++ packages/server/src/scheme/article.schema.ts
@@ -50,6 +50,9 @@ export class Article extends Document {
   @Prop()
   copyright?: string;
 
+  @Prop({ default: '' })
+  cover?: string;
+
   @Prop()
   lastVisitedTime: Date;